ref: Move code into tracing integration
