+++ packages/apm/src/integrations/index.ts
@@ -1,3 +1,2 @@
 export { Express } from './express';
-export { Tracing } from './tracing';
-export { TransactionActivity, TransactionActivityHandlers } from './transactionactivity';
+export { Tracing, TracingHandlers } from './tracing';
